screen vars in their own file
